Implemented Johannes' suggestions
