+++ lib/api.js
@@ -1,11 +1,11 @@
 "use strict";
 
-const moment = require("moment");
-const { ClockodoApi } = require("./lib.js");
+const { ClockodoLib } = require("./lib.js");
 
 const clockodoApi = Symbol("api");
-const UNBILLABLE = 0;
-const BILLABLE = 2;
+const ENTRY_UNBILLABLE = 0;
+const ENTRY_NOT_BILLED = 1;
+const ENTRY_BILLED = 2;
 const REQUIRED_PARAMS_GET_TASK_DURATION = [
     "taskCustomerId",
     "taskProjectId",
@@ -28,7 +28,7 @@ class Clockodo {
             throw new Error("No Clockodo 'apiKey' set");
         }
 
-        this[clockodoApi] = new ClockodoApi(user, apiKey);
+        this[clockodoApi] = new ClockodoLib({ user, apiKey });
     }
 
     getUser(id) {
@@ -62,16 +62,7 @@ class Clockodo {
     getEntries(params) {
         _checkRequired(params, REQUIRED_PARAMS_GET_ENTRIES);
 
-        let { begin, end } = params;
-
-        begin = moment(begin).format("YYYY-MM-DD hh:mm:ss");
-        end = moment(end).format("YYYY-MM-DD hh:mm:ss");
-
-        return this[clockodoApi].get("/entries", {
-            ...params,
-            begin,
-            end,
-        });
+        return this[clockodoApi].get("/entries", params);
     }
 
     getTaskDuration(params) {
@@ -87,16 +78,7 @@ class Clockodo {
     getEntryGroups(params) {
         _checkRequired(params, REQUIRED_PARAMS_GET_ENTRYGROUPS);
 
-        let { begin, end } = params;
-
-        begin = moment(begin).format("YYYY-MM-DD hh:mm:ss");
-        end = moment(end).format("YYYY-MM-DD hh:mm:ss");
-
-        return this[clockodoApi].get("/entrygroups", {
-            ...params,
-            begin,
-            end,
-        });
+        return this[clockodoApi].get("/entrygroups", params);
     }
 
     getUserReport(id, params) {
@@ -128,34 +110,18 @@ class Clockodo {
     getClockStatus() {
         return this[clockodoApi].get("/clock/update");
     }
-
-    // Convienece Methods
-    getBillableEntriesBetween(begin, end) {
-        return this.getEntries(begin, end, { filterBillable: BILLABLE });
-    }
-    getNonBillableEntriesBetween(begin, end) {
-        return this.getEntries(begin, end, { filterBillable: UNBILLABLE });
-    }
-    getUserEntriesBetween(userId, begin, end) {
-        return this.getEntries(begin, end, { filterUserId: userId });
-    }
-    getCustomerEntriesBetween(customerId, begin, end) {
-        return this.getEntries(begin, end, { filterCustomerId: customerId });
-    }
-    getServiceEntriesBetween(serviceId, begin, end) {
-        return this.getEntries(begin, end, { filterServiceId: serviceId });
-    }
-    getProjectEntriesBetween(projectId, begin, end) {
-        return this.getEntries(begin, end, { filterProjectId: projectId });
-    }
 }
 
 function _checkRequired(params, requiredList) {
     const missingParamName = requiredList.find(paramName => paramName in params === false);
 
     if (typeof missingParamName !== "undefined") {
-        throw new Error(`Missing Required Parameter "${ missingParamName }"`);
+        throw new Error(`Missing required parameter "${ missingParamName }"`);
     }
 }
+
+Clockodo.ENTRY_BILLED = ENTRY_BILLED;
+Clockodo.ENTRY_NOT_BILLED = ENTRY_NOT_BILLED;
+Clockodo.ENTRY_UNBILLABLE = ENTRY_UNBILLABLE;
 
 module.exports = Clockodo;
