+++ lib/lib.js
@@ -26,8 +26,8 @@ const paramMapping = {
     calcHardBudgetRevenues: "calc_also_revenues_for_projects_with_hard_budget",
 };
 
-class ClockodoApi {
-    constructor(user, apiKey) {
+class ClockodoLib {
+    constructor({ user, apiKey }) {
         this[axiosClient] = axios.create({
             baseURL: ENDPOINT,
             headers: {
@@ -44,7 +44,7 @@ class ClockodoApi {
             },
         });
 
-        return deepMapKeys(response.data, _camelize);
+        return deepMapKeys(response.data, key => camelCase(key));
     }
 }
 
@@ -60,10 +60,6 @@ function mapParams(userParams) {
     return apiParams;
 }
 
-function _camelize(key) {
-    return camelCase(key);
-}
-
 module.exports = {
-    ClockodoApi,
+    ClockodoLib,
 };
